--- a/mul_arm64.h
+++ b/mul_arm64.h
@@ -52,9 +52,9 @@
 	UMULH R4, R5, R26 \
 	MUL R4, R6, R0 \
 	ADDS R0, R26 \
-	UMULH R4, R6, R27 \
+	UMULH R4, R6, R5 \
 	MUL R4, R7, R0 \
-	ADCS R0, R27 \
+	ADCS R0, R5 \
 	UMULH R4, R7, R6 \
 	MUL R4, R8, R0 \
 	ADCS R0, R6 \
@@ -62,7 +62,7 @@
 	ADCS ZR, c7 \
 	ADDS R1, c3 \
 	ADCS R26, c4 \
-	ADCS R27, c5 \
+	ADCS R5, c5 \
 	ADCS R6, c6 \
 	ADCS  ZR, c7
 
@@ -121,6 +121,7 @@
 	ADCS R16, R24 \
 	ADCS  ZR, R0 \
 	\
+	MOVD ·p2+0(SB), R5 \ // Restore R5
 	MOVD ·p2+8(SB), R6 \ // Restore R6
 	\ // Our output is R21:R22:R23:R24. Reduce mod p if necessary.
 	SUBS R5, R21, R10 \