Resolves #46.

This removes all usage of `R27` and `R29` from `gfpMul`.

This fix introduces 3 additional `MOVD` instructions (1 for the removal of `R29` and 2 for the removal of `R27`).

I couldn't figure out any less costly way of removing the usage of the registers (as all of the available registers are used).

I definitely don't love the style of the change (especially from the last commit where the additional arguments to `mul` were added)... So any suggestions there would be especially appreciated.