@StephenButtolph 2025-05-12
Probably should have included benchmarks:

Running locally on my laptop:
```
goos: darwin
goarch: arm64
pkg: github.com/cloudflare/bn256
cpu: Apple M2 Max
                    │ master.bench │       fix-arm64-crash.bench        │
                    │    sec/op    │   sec/op     vs base               │
G1-12                  72.84µ ± 4%   73.67µ ± 3%       ~ (p=0.912 n=10)
G2-12                  219.8µ ± 3%   224.4µ ± 4%       ~ (p=0.218 n=10)
GT-12                  553.0µ ± 4%   549.5µ ± 5%       ~ (p=0.853 n=10)
Pairing-12             696.9µ ± 0%   699.6µ ± 0%       ~ (p=0.089 n=10)
HashG1Size8bytes-12    25.20µ ± 1%   25.47µ ± 1%  +1.09% (p=0.029 n=10)
HashG1Size1k-12        19.55µ ± 1%   19.66µ ± 1%  +0.56% (p=0.015 n=10)
HashG1Size8k-12        27.90µ ± 1%   28.17µ ± 1%  +0.98% (p=0.000 n=10)
geomean                97.67µ        98.45µ       +0.80%

                    │ master.bench │         fix-arm64-crash.bench         │
                    │     B/op     │     B/op      vs base                 │
G1-12                   160.0 ± 0%     160.0 ± 0%       ~ (p=1.000 n=10) ¹
G2-12                   256.0 ± 0%     256.0 ± 0%       ~ (p=1.000 n=10) ¹
GT-12                   416.0 ± 0%     416.0 ± 0%       ~ (p=1.000 n=10) ¹
Pairing-12            38.25Ki ± 0%   38.25Ki ± 0%       ~ (p=1.000 n=10) ¹
HashG1Size8bytes-12   1.938Ki ± 0%   1.938Ki ± 0%       ~ (p=1.000 n=10) ¹
HashG1Size1k-12       1.907Ki ± 0%   1.907Ki ± 0%       ~ (p=1.000 n=10) ¹
HashG1Size8k-12       1.938Ki ± 0%   1.938Ki ± 0%       ~ (p=1.000 n=10) ¹
geomean               1.234Ki        1.234Ki       +0.00%
¹ all samples are equal

                    │ master.bench │        fix-arm64-crash.bench        │
                    │  allocs/op   │ allocs/op   vs base                 │
G1-12                   2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=10) ¹
G2-12                   1.000 ± 0%   1.000 ± 0%       ~ (p=1.000 n=10) ¹
GT-12                   2.000 ± 0%   2.000 ± 0%       ~ (p=1.000 n=10) ¹
Pairing-12              352.0 ± 0%   352.0 ± 0%       ~ (p=1.000 n=10) ¹
HashG1Size8bytes-12     29.00 ± 0%   29.00 ± 0%       ~ (p=1.000 n=10) ¹
HashG1Size1k-12         28.00 ± 0%   28.00 ± 0%       ~ (p=1.000 n=10) ¹
HashG1Size8k-12         29.00 ± 0%   29.00 ± 0%       ~ (p=1.000 n=10) ¹
geomean                 11.87        11.87       +0.00%
¹ all samples are equal

                    │ master.bench │        fix-arm64-crash.bench        │
                    │     B/s      │     B/s       vs base               │
HashG1Size8bytes-12   312.5Ki ± 0%   302.7Ki ± 3%       ~ (p=0.057 n=10)
HashG1Size1k-12       49.96Mi ± 1%   49.68Mi ± 1%  -0.55% (p=0.016 n=10)
HashG1Size8k-12       280.0Mi ± 1%   277.3Mi ± 1%  -0.97% (p=0.000 n=10)
geomean               16.22Mi        15.97Mi       -1.56%
```

As expected (tbh probably within random perturbation) the new code is slightly slower.