@armfazh 2025-05-19
Closing since I found a way in #48 to remove reserved registers without using stack based on your changes.
Also I could not reproduce the test, test passes even though the reserved registers are used.